Developer branch

+++ TAABP.Application/Services/RoomService.cs
@@ -3,6 +3,7 @@ using TAABP.Application.Exceptions;
 using TAABP.Application.Profile.RoomMapping;
 using TAABP.Application.RepositoryInterfaces;
 using TAABP.Application.ServiceInterfaces;
+using TAABP.Core;
 
 namespace TAABP.Application.Services
 {
@@ -11,26 +12,30 @@ namespace TAABP.Application.Services
         private readonly IRoomRepository _roomRepository;
         private readonly IRoomMapper _roomMapper;
         private readonly IHotelRepository _hotelRepository;
-        public RoomService(IRoomRepository roomRepository, IRoomMapper roomMapper, IHotelRepository hotelRepository)
+        private readonly IUserService _userService;
+        public RoomService(IRoomRepository roomRepository, IRoomMapper roomMapper,
+            IHotelRepository hotelRepository, IUserService userService)
         {
             _roomRepository = roomRepository;
             _roomMapper = roomMapper;
             _hotelRepository = hotelRepository;
+            _userService = userService;
         }
 
-        public async Task<RoomDto> GetRoomAsync(int id)
+        public async Task<RoomDto> GetRoomByIdAsync(int hotelId, int id)
         {
+            var hotel = await _hotelRepository.GetHotelByIdAsync(hotelId);
             var room = await _roomRepository.GetRoomByIdAsync(id);
-            if(room == null)
+            if(hotel == null || room == null  || room.HotelId != hotelId)
             {
-                throw new EntityNotFoundException("Room not found");
+                throw new EntityNotFoundException("Hotel Or Room not found");
             }
             return _roomMapper.RoomToRoomDto(room);
         }
 
         public async Task<List<RoomDto>> GetRoomsAsync(int hotelId)
         {
-            if (await _hotelRepository.GetHotelAsync(hotelId) == null)
+            if (await _hotelRepository.GetHotelByIdAsync(hotelId) == null)
             {
                 throw new EntityNotFoundException("Hotel not found");
             }
@@ -38,49 +43,49 @@ namespace TAABP.Application.Services
             return rooms.Select(r => _roomMapper.RoomToRoomDto(r)).ToList();
         }
 
-        public async Task CreateRoomAsync(RoomDto roomDto)
+        public async Task<int> CreateRoomAsync(RoomDto roomDto)
         {
-            var hotel = await _hotelRepository.GetHotelAsync(roomDto.HotelId);
+            var hotel = await _hotelRepository.GetHotelByIdAsync(roomDto.HotelId);
             if (hotel == null)
             {
                 throw new EntityNotFoundException("Hotel not found");
             }
-
-            var room = _roomMapper.RoomDtoToRoom(roomDto);
+            var room = new Room();
+            _roomMapper.RoomDtoToRoom(roomDto, room);
             room.CreatedAt = DateTime.Now;
-            room.CreatedBy = "System";
+            room.CreatedBy = await _userService.GetCurrentUsernameAsync();
             await _roomRepository.CreateRoomAsync(room);
+            await _hotelRepository.IncrementNumberOfRoomsAsync(roomDto.HotelId);
+            return room.RoomId;
         }
 
         public async Task UpdateRoomAsync(RoomDto roomDto)
         {
-            var hotel = await _hotelRepository.GetHotelAsync(roomDto.HotelId);
-            if (hotel == null)
-            {
-                throw new EntityNotFoundException("Hotel not found");
-            }
+            var hotel = await _hotelRepository.GetHotelByIdAsync(roomDto.HotelId);
             var room = await _roomRepository.GetRoomByIdAsync(roomDto.RoomId);
-            if (room == null)
+            if(hotel == null || room == null || room.HotelId != roomDto.HotelId)
             {
-                throw new EntityNotFoundException("Room not found");
+                throw new EntityNotFoundException("Hotel Or Room not found");
             }
-            room = _roomMapper.RoomDtoToRoom(roomDto);
+            _roomMapper.RoomDtoToRoom(roomDto, room);
             room.UpdatedAt = DateTime.Now;
-            room.UpdatedBy = "System";
+            room.UpdatedBy = await _userService.GetCurrentUsernameAsync();
             await _roomRepository.UpdateRoomAsync(room);
         }
 
-        public async Task DeleteRoomAsync(int id)
+        public async Task DeleteRoomAsync(int hotelId , int id)
         {
+            var hotel = await _hotelRepository.GetHotelByIdAsync(hotelId);
             var room = await _roomRepository.GetRoomByIdAsync(id);
-            if (room == null)
+            if (hotel == null || room == null || room.HotelId != hotelId)
             {
-                throw new EntityNotFoundException("Room not found");
+                throw new EntityNotFoundException("Hotel Or Room not found");
             }
             await _roomRepository.DeleteRoomAsync(room);
+            await _hotelRepository.DecrementNumberOfRoomsAsync(hotelId);
         }
 
-        public async Task CreateRoomImageAsync(RoomImageDto roomImageDto)
+        public async Task<int> CreateRoomImageAsync(RoomImageDto roomImageDto)
         {
             var room = await _roomRepository.GetRoomByIdAsync(roomImageDto.RoomId);
             if (room == null)
@@ -89,24 +94,16 @@ namespace TAABP.Application.Services
             }
             var roomImage = _roomMapper.RoomImageDtoToRoomImage(roomImageDto);
             await _roomRepository.CreateRoomImageAsync(roomImage);
+            return roomImage.RoomImageId;
         }
 
-        public async Task DeleteRoomImageAsync(int id)
+        public async Task<RoomImageDto> GetRoomImageByIdAsync(int roomId, int id)
         {
-            var roomImage = await _roomRepository.GetRoomImageAsync(id);
-            if (roomImage == null)
+            var roomImage = await _roomRepository.GetRoomImageByIdAsync(id);
+            var room = await _roomRepository.GetRoomByIdAsync(roomId);
+            if (roomImage == null || room == null || room.RoomId != roomImage.RoomId)
             {
-                throw new EntityNotFoundException("Room image not found");
-            }
-            await _roomRepository.DeleteRoomImageAsync(roomImage);
-        }
-
-        public async Task<RoomImageDto> GetRoomImageAsync(int id)
-        {
-            var roomImage = await _roomRepository.GetRoomImageAsync(id);
-            if (roomImage == null)
-            {
-                throw new EntityNotFoundException("Room image not found");
+                throw new EntityNotFoundException("Room or Room image not found");
             }
             return _roomMapper.RoomImageToRoomImageDto(roomImage);
         }
@@ -120,6 +117,17 @@ namespace TAABP.Application.Services
             }
             var roomImages = await _roomRepository.GetRoomImagesAsync(roomId);
             return roomImages.Select(ri => _roomMapper.RoomImageToRoomImageDto(ri)).ToList();
+        }
+
+        public async Task DeleteRoomImageAsync(int roomId, int id)
+        {
+            var roomImage = await _roomRepository.GetRoomImageByIdAsync(id);
+            var room = await _roomRepository.GetRoomByIdAsync(roomId);
+            if (roomImage == null || room == null || room.RoomId != roomImage.RoomId)
+            {
+                throw new EntityNotFoundException("Room or Room image not found");
+            }
+            await _roomRepository.DeleteRoomImageAsync(roomImage);
         }
     }
 }
